updates / changes: - the main.dart file doesn't contain the home view anymore it just loads the home page that is now in the home_view.dart file
- the bottom navigation bar has the 3 main functions (for now) and when each button is pressed it navigates to a view page that will display what it holds (empty right now)

- the app bar goes away when the home page is left - I'm looking at different ways to fix that I don't think it will be hard

- there isn't an app bar on the homepage I know, but that's easy to implement and I wanted to push these changes for now

- theres obviously no styling right now, but I'm working to get the flow of logic in check before making it look pretty

thinking:
For the home page we just fill it up with bullshit and let the bottom navigation bar navigate to all of the core features (i.e. the things we talked to jacob and the TA dude about). It'd be best to keep it simple by showing the 3 main things - pairing, favoriting, and a profile.

Entering food allergies and shit will be a good thing to showcase in the profile section, but having it affect the pairing recommendations might be waaaaaaaaay over our heads - I'm not sure if he will be expecting to have that fully functional but it doesn't matter right now since we only need to get the pairing function working for Thursday.

More on potential shit for the homepage we could make it look pretty with a nice logo at the top and maybe a slogan and put some fun shit like this to fill it up:
"Pairings We Love" section that just suggests a random pairing

"Explore Restaurants Near You" that just pulls up a map or links to google maps

"Whats new" maybe a link to new news in the alcohol industry or something

Those are just some ideas, filling the home with buttons of our core functionalities I feel like would look a little silly, the nav bar can do all that work. Luckily we dont really have to worry about that now - gotta work on the pairings page